+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not yet released
 
 * Add support for comparison across forks and better detect dev versions ([#19](https://github.com/pyrech/composer-changelogs/pull/19))
+* Add autoloading of classes required to make the plugin always working ([#22](https://github.com/pyrech/composer-changelogs/pull/22))
 
 ## 1.2 (2015-10-22)